Move stats to column chunk data
